gpu_server/requirements.txt
# Core FastAPI dependencies
fastapi==0.116.2
uvicorn[standard]==0.30.6
pydantic==2.11.9
python-multipart==0.0.9
python-dotenv>=1.0.0


# GeoCLIP and image processing
geoclip>=1.0.0
pillow>=10.0.0
pillow-heif>=0.16.0
numpy>=1.24.0

# ML and image processing - GPU-enabled versions
--index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0

# Geocoding
reverse_geocoder>=1.5.1

# HTTP client for image downloads
aiohttp>=3.9.0
